+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Ensure Topic Operator using Kafka Streams state store can start up successfully 
 * Update Cruise Control to 2.5.89
 * Remove TLS sidecar from Cruise Control pod. Cruise Control is now configured to not using ZooKeeper, so the TLS sidecar is not needed anymore.
+* Allow Cruise Control topic names to be configured
 
 ### Changes, deprecations and removals